requirements.txt
absl-py==0.15.0
altair==4.1.0
anyio==3.3.2
appdirs==1.4.4
APScheduler==3.6.3
argcomplete==1.12.3
argon2-cffi==20.1.0
asn1crypto==1.4.0
astor==0.8.0
astroid==2.9.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.9.24
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.2.1
cryptography==2.5
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
cytoolz==0.9.0.1
dask==2.1.0
dataclasses==0.8
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.6
dill==0.2.9
dlib==19.21.0
docker-pycreds==0.4.0
dominate==2.3.5
easydict==1.9
entrypoints==0.3
face-alignment==1.3.5
filelock==3.3.1
flake8==3.9.2
gast==0.2.2
gdown==4.2.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.43.0
h5py==2.10.0
helpdev==0.6.10
idna==2.10
imageio==2.5.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.8.1
imutils==0.5.3
in-place==0.5.0
ipdb==0.13.9
ipykernel==6.3.1
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==5.9.3
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.0.0
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupytext==1.11.5
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==23.2.1
kiwisolver==1.1.0
knockknock==0.1.8.1
lazy-object-proxy==1.6.0
llvmlite==0.37.0
lmdb==1.3.0
lpips==0.1.4
lxml==4.6.3
Mako==1.1.2
Markdown==3.3.4
markdown-it-py==1.1.0
MarkupSafe==2.0.1
matplotlib==3.1.0
matplotlib-inline==0.1.2
matrix-client==0.4.0
mccabe==0.6.1
mdit-py-plugins==0.2.8
mistune==0.8.4
mit-semseg==1.0.0
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.2
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.3
notebook==6.4.4
numba==0.54.1
numpy==1.20.3
nvidia-dali-cuda110==1.10.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.0.25
opt-einsum==3.2.1
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==21.2.2
platformdirs==2.4.0
premailer==3.10.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.22
protobuf==3.17.3
psutil==5.6.3
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycuda==2019.1.2
pydeck==0.7.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.3.0
pylint==2.12.2
pynvim==0.4.3
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyQt5==5.13.0
PyQt5_sip==4.19.18
pyrsistent==0.17.3
PySide2==5.13.0
PySocks==1.7.1
pyspng==0.1.0
python-dateutil==2.8.2
python-telegram-bot==13.7
pytools==2020.1
pytorch-fid==0.2.1
pytz==2021.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==22.1.0
qdarkgraystyle==1.0.2
QDarkStyle==2.7
qtconsole==4.5.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.24.2
scikit-video==1.1.11
scipy==1.2.1
SecretStorage==3.3.1
Send2Trash==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==52.0.0.post20210125
shiboken2==5.13.0
shortuuid==1.0.9
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.2.1
streambook==0.2
streamlit==0.88.0
streamlit-drawable-canvas==0.8.0
svgpathtools==1.4.1
svgwrite==1.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.9.0
torch==1.10.2
torch-tb-profiler==0.3.1
torchaudio==0.10.2
torchdiffeq==0.0.1
torchvision==0.11.3
tornado==6.1
tqdm==4.32.1
traitlets==5.0.5
twilio==7.2.0
typed-ast==1.4.3
typer==0.4.0
typing_extensions==4.4.0
tzlocal==3.0
urllib3==1.26.7
validators==0.18.2
wandb==0.12.17
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.0.1
wheel==0.37.0
widgetsnbextension==3.4.2
wrapt==1.13.3
yacs==0.1.8
yagmail==0.14.260
zipp==3.5.0
